third_party/blink/web_tests/platform/generic/html5lib/generated/run-tests11-write-expected.txt

CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
../resources/tests11.dat: PASS
